cleanup code
remove code duplicates (use MoneyRoundingMode trait)
